+++ src/components/ui/Button.jsx
@@ -1,4 +1,3 @@
-// Button Component (button.jsx)
 import React from "react";
 import { cn } from "@/lib/utils";
